Move main layout module  to new lib layout
