+++ projects/storefrontlib/src/layout/config/default-layout-config.ts
