+++ projects/storefrontlib/src/layout/config/layout-config.ts
