+++ projects/storefrontlib/src/layout/index.ts
@@ -1,1 +1,7 @@
+export * from './breakpoint/breakpoint.service';
+export * from './config/default-layout-config';
+export * from './config/layout-config';
 export * from './header/index';
+export * from './layout.module';
+export * from './main/main.module';
+export * from './main/storefront.component';
