+++ projects/storefrontlib/src/layout/layout.module.ts
@@ -1,8 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Config, ConfigModule } from '@spartacus/core';
-import { OutletRefModule } from '../../outlet/index';
-import { StyleRefModule } from '../../outlet/style-ref/style-ref.module';
-import { BreakpointService } from './breakpoint/breakpoint.service';
+import { OutletRefModule } from '../lib/outlet/index';
+import { StyleRefModule } from '../lib/outlet/style-ref/style-ref.module';
 import { defaultLayoutConfig } from './config/default-layout-config';
 import { LayoutConfig } from './config/layout-config';
 import { MainModule } from './main/main.module';
@@ -15,10 +14,7 @@ const layoutModules = [OutletRefModule, StyleRefModule];
     ...layoutModules,
     ConfigModule.withConfig(defaultLayoutConfig),
   ],
-  providers: [
-    { provide: LayoutConfig, useExisting: Config },
-    BreakpointService,
-  ],
+  providers: [{ provide: LayoutConfig, useExisting: Config }],
   exports: [MainModule, ...layoutModules],
 })
 export class LayoutModule {}
