+++ projects/storefrontlib/src/lib/ui/index.ts
@@ -1,6 +1,3 @@
-export * from './ui.module';
-export * from './ui-framework/ui-framework.module';
-
 export * from './components/index';
 export * from './pages/index';
-export * from './layout/index';
+export * from './ui.module';
