--- projects/storefrontlib/src/lib/ui/layout/index.ts
@@ -1,7 +1,0 @@
-export * from './breakpoint/breakpoint.service';
-export * from './config/default-layout-config';
-export * from './config/layout-config';
-export * from './layout.module';
-export * from './main/main.module';
-// tmp: we remove those once we render them all with page layout component
-export * from './main/storefront.component';
