--- projects/storefrontlib/src/lib/ui/ui-framework/ui-framework.module.ts
@@ -1,10 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-
-@NgModule({
-  imports: [CommonModule, BrowserAnimationsModule],
-  declarations: [],
-})
-export class UiFrameworkModule {}
