+++ projects/storefrontlib/src/lib/ui/ui.module.ts
@@ -1,7 +1,6 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-
-import { LayoutModule } from './layout/layout.module';
+import { NgModule } from '@angular/core';
+import { LayoutModule } from '../../layout/layout.module';
 import { PagesModule } from './pages/pages.module';
 
 @NgModule({
